Revert "combine api context and loading"
This reverts commit b8fbf24486b8122b7d336de13d8d375ebaf7518b.
